requirements.txt
Flask
sympy
matplotlib
fpdf
pillow
pytesseract